Выпоолнено задание первой недели
