+++ package.json
@@ -8,8 +8,9 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
-    "web-vitals": "^2.1.4"
+    "react-scripts": "^2.1.3",
+    "web-vitals": "^2.1.4",
+    "yarn": "1.22.19"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -34,5 +35,14 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
-  }
+  },
+  "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
+  "main": "index.js",
+  "repository": "git+https://github.com/Barsibarsik/loft-taxi.git",
+  "author": "Margarita Samburskaia <barsibarsik@mail.ru>",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Barsibarsik/loft-taxi/issues"
+  },
+  "homepage": "https://github.com/Barsibarsik/loft-taxi#readme"
 }
