--- src/Components/Map.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import { OrderMake } from './OrderMake';
-
-export const Map = () => {
-    return <div>
-                <div>Map</div>
-                <div><OrderMake /></div>
-            </div>
-}
