--- src/Components/OrderDone.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-export const OrderDone = () => {
-    return <div>
-                <h1 className="title">Заказ размещён</h1>
-                <p>Ваше такси уже едет к вам. Прибудет приблизительно через 10 минут.</p>
-                <button>Сделать новый заказ</button>
-            </div>
-}
