--- src/Components/OrderMake.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-
-export const OrderMake = () => {
-    return <form>
-                <h1 className="title">Форма заказа</h1>
-                <input type="submit" value="Заказать" />
-            </form>
-}
