+++ src/Components/order-make/OrderMake.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+
+export default function OrderMake() {
+     return <form> 
+                <h1 className="title">Форма заказа</h1>
+                <input type="submit" value="Заказать" />
+            </form>
+}
