+++ src/Components/profile-done/ProfileDone.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const ProfileDone = () => {
+    return <div>
+                <h1 className="title">Профиль</h1>
+                <h4>Платёжные данные обновлены. Теперь вы можете заказывать такси.</h4>
+                <button>Перейти на карту</button>
+            </div>
+}
